Remove the /health endpoint
