+++ docs/deployment.rst
@@ -103,13 +103,13 @@ Kafka Connect
 =============
 When deploying in conjunction with Kafka Connect, the Avro schemas need to be pre-registered with the `Schema Registry <https://docs.confluent.io/3.3.0/schema-registry/docs>`_. Mappings that produce records for a Kafka sink operating in ``confluent`` mode need have their ``confluent_id`` property configured with the identifier of the schema in the registry. (This identifier is normally a simple integer.)
 
-Service endpoints
-=================
+Service endpoint
+================
 
-To support rolling updates and scaling without losing data, Divolte exposes an endpoint that allows save shutdown of the docker container. Currently the `/health` endpoint is tested on Kubernetes. The main issue with Kubernetes is, even it has send a signal to stop the container, the container will remain in the pool of the load balancer. This will cause the divolte container to receive new connections, which is not something that we want when initializing a graceful shutdown procedure. Instead Divolte should clear all the buffers by flushing them to persistent storage. The sequence:
+To support rolling updates and scaling without losing data, Divolte exposes an endpoint that allows graceful shutdown of the docker container. Currently the `/ping` endpoint is tested on Kubernetes. The main issue with Kubernetes is, even it has send a signal to stop the container, the container will remain in the pool of the load balancer. This will cause the divolte container to receive new connections, which is not something that we want when initializing a graceful shutdown procedure. Instead Divolte should clear all the buffers by flushing them to persistent storage. The sequence:
 
 - Kubernetes sends a SIGTERM to Divolte to let it know that it wants to shutdown the container.
-- Making the container unhealthy by replying with HTTP 503 over the `/health` endpoint, which will cause the container to be removed from the pool of the load balancer.
+- Making the container unhealthy by replying with HTTP 503 over the `/ping` endpoint, which will cause the container to be removed from the pool of the load balancer.
 - The `divolte.global.server.shutdown_wait_period_mills` property should be at least twice the kubernetes liveness probe interval (10 seconds by default).
 - The container is removed from the loadbalancer pool and divolte will wait `divolte.global.server.shutdown_grace_period_mills` to handle and close the open connections.
 - Undertow will shut down, and all the sinks are flushed and closed.
